fix: Possible bug when DST kicks in on selected zone
